// src/app/generated/blueprint.api/model/competencyElement.ts
/*
 Copyright 2026 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Blueprint API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompetencyElement {
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    competencyFrameworkId?: string;
    elementIdentifier?: string | null;
    elementType?: string | null;
    name?: string | null;
    description?: string | null;
    parentId?: string | null;
    children?: Array<CompetencyElement> | null;
}